feat: Add JobCode loading on Add Job Spec page
Clicking on a new job code on the homepage will direct the user
to the Add Job Spec page, and then load information for that
Job Code that is stored in the Pay Levels table.

This update made it so the Add Job Spec page loads the Job Code
information.